Consistent code folder/AAR naming, fix spotless issues
