--- code/android-campaignclassic-library/release.gradle
@@ -1,115 +1,0 @@
-/**
- * Copyright 2022 Adobe. All rights reserved.
- * This file is licensed to you under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License. You may obtain a copy
- * of the License at http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under
- * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
- * OF ANY KIND, either express or implied. See the License for the specific language
- * governing permissions and limitations under the License.
- */
-apply plugin: 'signing'
-apply plugin: 'maven-publish'
-apply plugin: 'org.jetbrains.dokka'
-
-
-def isReleaseBuild() {
-    return hasProperty('release')
-}
-
-def groupIdForPublish = 'com.adobe.marketing.mobile'
-version = isReleaseBuild() ? rootProject.campaignClassicExtensionVersion : rootProject.campaignClassicExtensionVersion+"-SNAPSHOT"
-
-dokkaJavadoc.configure {
-    dokkaSourceSets {
-        named("main") {
-            noAndroidSdkLink.set(false)
-            perPackageOption {
-                matchingRegex.set(".*\\.internal.*") // proper setting
-                suppress.set(true)
-            }
-        }
-
-        named("phone") {
-            noAndroidSdkLink.set(false)
-            perPackageOption {
-                matchingRegex.set(".*\\.internal.*") // proper setting
-                suppress.set(true)
-            }
-        }
-    }
-}
-
-task javadocJar(type: Jar, dependsOn: dokkaJavadoc) {
-    archiveClassifier.set('javadoc')
-    from dokkaJavadoc.outputDirectory
-}
-
-publishing {
-    publications {
-        release(MavenPublication) {
-            groupId = groupIdForPublish
-            artifactId = rootProject.campaignClassicExtensionName
-            version = version
-            artifact("$buildDir/outputs/aar/${rootProject.campaignClassicExtensionAARName}")
-            artifact javadocJar
-            pom {
-                name = rootProject.mavenRepoName
-                description = rootProject.mavenRepoDescription
-                url = 'https://developer.adobe.com/client-sdks'
-                licenses {
-                    license {
-                        name = 'Adobe Proprietary'
-                    }
-                }
-                developers {
-                    developer {
-                        id = 'adobe'
-                        name = 'adobe'
-                        email = 'adobe-mobile-testing@adobe.com'
-                    }
-                }
-                scm {
-                    connection = 'scm:git:github.com//adobe/aepsdk-campaignclassic-android.git'
-                    developerConnection = 'scm:git:ssh://github.com//adobe/aepsdk-campaignclassic-android.git'
-                    url = 'https://github.com/adobe/aepsdk-campaignclassic-android'
-                }
-                withXml {
-                    def dependenciesNode = asNode().appendNode('dependencies')
-
-                    def coreDependencyNode = dependenciesNode.appendNode('dependency')
-                    coreDependencyNode.appendNode('groupId', 'com.adobe.marketing.mobile')
-                    coreDependencyNode.appendNode('artifactId', 'core')
-                    coreDependencyNode.appendNode('version', mavenCoreVersion)
-                }
-            }
-        }
-    }
-    repositories {
-        maven {
-            name = "sonatype"
-            url = version.endsWith('SNAPSHOT') ? "https://oss.sonatype.org/content/repositories/snapshots/" : "https://oss.sonatype.org/service/local/staging/deploy/maven2/"
-            credentials {
-                username = System.getenv("SONATYPE_USERNAME")
-                password = System.getenv("SONATYPE_PASSWORD")
-            }
-        }
-    }
-}
-tasks.withType(Javadoc).all { enabled = false }
-publish.dependsOn('assemblePhone')
-
-ext."signing.gnupg.executable"="gpg"
-ext."signing.gnupg.keyName"=System.getenv("GPG_KEY_ID")
-ext."signing.gnupg.passphrase"=System.getenv("GPG_PASSPHRASE")
-
-signing {
-    useGpgCmd()
-    required {
-        tasks.withType(PublishToMavenRepository).find {
-            gradle.taskGraph.hasTask it
-        }
-    }
-    sign publishing.publications
-}
