+++ code/campaignclassic/.gitignore
