+++ code/campaignclassic/proguard-rules.pro
