+++ code/campaignclassic/src/androidTest/java/com/adobe/marketing/mobile/SDKHelper.java
@@ -1,5 +1,5 @@
 /*
-  Copyright $YEAR Adobe. All rights reserved.
+  Copyright 2022 Adobe. All rights reserved.
   This file is licensed to you under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License. You may obtain a copy
   of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -8,3 +8,10 @@
   OF ANY KIND, either express or implied. See the License for the specific language
   governing permissions and limitations under the License.
 */
+package com.adobe.marketing.mobile;
+
+public class SDKHelper {
+    public static void resetSDK() {
+        MobileCore.resetSDK();
+    }
+}
