+++ code/campaignclassic/src/main/AndroidManifest.xml
