+++ code/campaignclassic/src/main/res/values/strings.xml
