+++ code/gradle.properties
@@ -14,15 +14,16 @@ org.gradle.jvmargs=-Xmx2048m
 android.injected.testOnly = false
 org.gradle.configureondemand=false
 
-moduleProjectName=android-campaignclassic-library
+moduleProjectName=campaignclassic
+moduleName=campaignclassic
+moduleVersion=2.0.0
+moduleAARName=campaignclassic-phone-release.aar
 
 #Maven artifact
-campaignClassicExtensionName = campaignclassic
 mavenRepoName=AdobeMobileCampaignClassicSdk
 mavenRepoDescription=Android Campaign Classic Extension for Adobe Mobile Marketing
 mavenUploadDryRunFlag=false
-campaignClassicExtensionVersion=2.0.0
-campaignClassicExtensionAARName=android-campaignclassic-library-phone-release.aar
+mavenCoreVersion=2.0.0
 
 android.disableAutomaticComponentCreation=true
 android.useAndroidX=true
