+++ code/settings.gradle
@@ -9,6 +9,22 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
- include ':android-campaignclassic-library', ':testapp'
+pluginManagement {
+ repositories {
+  gradlePluginPortal()
+  google()
+  mavenCentral()
+ }
+}
+dependencyResolutionManagement {
+ repositoriesMode.set(RepositoriesMode.FAIL_ON_PROJECT_REPOS)
+ repositories {
+  google()
+  mavenCentral()
+  maven { url "https://oss.sonatype.org/content/repositories/snapshots/" }
+ }
+}
 
-rootProject.name = "aepsdk-campaignclassic-android"
+ include ':campaignclassic', ':testapp'
+
+rootProject.name = "campaignclassic-sdk"
